Conversion of projections objects

These functions convert projections objects into other classes.

# S3 method for projections
as.matrix(x, ...)

# S3 method for projections
as.data.frame(x, ..., long = FALSE)

Arguments

x
An projections object, or an object to be converted as projections (see details).

...
Further arguments passed to other functions (no used).

long
A logical indicating if the output data.frame should be 'long', i.e. where a single column containing 'groups' is added in case of data computed on several groups.

See also