Subsetting 'projections' objects

Two functions can be used to subset projections objects. The operator "[" can be used as for matrices, using the syntax x[i,j] where 'i' is a subset of dates, and 'j' is a subset of simulations.

# S3 method for projections
[(x, i, j)

# S3 method for projections
subset(x, ..., from = NULL, to = NULL, sim = TRUE)

Arguments

x

An projections object, generated by the function project.

i

a subset of dates to retain

j

a subset of groups to retain

...

Further arguments passed to other methods (not used).

from

The starting date; data strictly before this date are discarded.

to

The ending date; data strictly after this date are discarded.

sim

(optional) The simulations to retained, indicated as subsets of the columns of x.

See also

Author

Thibaut Jombart thibautjombart@gmail.com