Summary for projections objects

This method summarises predicted epidemic trajectories contained in a projections object by days, deriving the mean, standard deviation, and user-specified quantiles for each day.

# S3 method for projections
summary(
  object,
  quantiles = c(0.025, 0.25, 0.5, 0.75, 0.975),
  mean = TRUE,
  sd = TRUE,
  min = TRUE,
  max = TRUE,
  ...
)

Arguments

object

A projections object to summarise

quantiles

A numeric vector indicating which quantiles should be computed; ignored if FALSE or of length 0

mean

a logical indicating of the mean should be computed

sd

a logical indicating of the standard deviation should be computed

min

a logical indicating of the minimum should be computed

max

a logical indicating of the maximum should be computed

...

only preesnt for compatibility with the generic

if (require(incidence)) i <- incidence::incidence(as.Date('2020-01-23')) si <- c(0.2, 0.5, 0.2, 0.1) R0 <- 2p <- project(x = i, si = si, R = R0, n_sim = 2, R_fix_within = TRUE, n_days = 10, model = "poisson" ) summary(p)

Author

Thibaut Jombart